+++ andhow-core/src/main/java/org/yarnandtail/andhow/load/MapLoader.java
@@ -7,84 +7,28 @@ import org.yarnandtail.andhow.api.*;
 /**
  * Loads values from a map.
  *
- * @author eeverman
  */
-public class MapLoader extends BaseLoader implements ReadLoader {
+public class MapLoader extends BaseMapLoader {
 
-	protected Map<?, ?> map;
+	protected Map<String, String> map;
 
-	protected boolean unknownPropertyAProblem = true;
-
-	public void setMap(Map<?, ?> map) {
+	public void setMap(Map<String, String> map) {
 		this.map = map == null ? null : new HashMap<>(map);
 	}
 
-	public Map<?, ?> getMap() {
+	public Map<String, String> getMap() {
 		return map;
 	}
 
 	@Override
-	public LoaderValues load(PropertyConfigurationInternal appConfigDef,
-                             ValidatedValuesWithContext existingValues) {
-
-		Map<?, ?> props = getMap();
-
-		if (props != null && !props.isEmpty()) {
-			ArrayList<ValidatedValue> values = new ArrayList();
-			ProblemList<Problem> problems = new ProblemList();
-
-			Set<?> keys = props.keySet();
-			for(Object key : keys) {
-				if (key != null) {
-					Object val = props.get(key);
-					String sVal = (val != null)?val.toString():null;
-
-					attemptToAdd(appConfigDef, values, problems, key.toString(), sVal);
-				}
-			}
-
-			values.trimToSize();
-			return new LoaderValues(this, values, problems);
-		} else {
-			return new LoaderValues(this);
-		}
-
-
-	}
-
-	@Override
-	public String getSpecificLoadDescription() {
-		return "Map";
-	}
-
-	@Override
-	public boolean isTrimmingRequiredForStringValues() {
-		return true;
-	}
-
-	// false is a reasonable default, since its only a special case (the command line loader) that
-	// would ever set this true.
-	@Override
-	public boolean isFlaggable() { return false; }
-
-	@Override
-	public void setUnknownPropertyAProblem(boolean isAProblem) {
-		unknownPropertyAProblem = isAProblem;
-	}
-
-	@Override
-	public boolean isUnknownPropertyAProblem() {
-		return unknownPropertyAProblem;
-	}
-
-	@Override
-	public String getLoaderType() {
-		return "Map";
-	}
-
-	@Override
-	public String getLoaderDialect() {
+	public LoaderValues load(final PropertyConfigurationInternal runtimeDef, final ValidatedValuesWithContext existingValues) {
 		return null;
+	}
+
+	@Override
+	public LoaderValues load(final PropertyConfigurationInternal runtimeDef,
+			final LoaderEnvironment environment, final ValidatedValuesWithContext existingValues) {
+		return load(runtimeDef, existingValues, map);
 	}
 
 	@Override
