+++ andhow-core/src/main/java/org/yarnandtail/andhow/load/std/StdSysPropLoader.java
@@ -1,7 +1,10 @@
 package org.yarnandtail.andhow.load.std;
 
 import java.util.*;
-import org.yarnandtail.andhow.api.StandardLoader;
+
+import org.yarnandtail.andhow.api.*;
+import org.yarnandtail.andhow.internal.PropertyConfigurationInternal;
+import org.yarnandtail.andhow.load.BaseMapLoader;
 import org.yarnandtail.andhow.load.MapLoader;
 
 /**
@@ -66,7 +69,7 @@ import org.yarnandtail.andhow.load.MapLoader;
  * be inserted into the load order via the
  * {@code AndHowConfiguration.insertLoaderBefore/After()}.
  */
-public class StdSysPropLoader extends MapLoader implements StandardLoader {
+public class StdSysPropLoader extends BaseMapLoader implements StandardLoader {
 	
 	/**
 	 * There is no reason to use the constructor in production application code
@@ -75,14 +78,16 @@ public class StdSysPropLoader extends MapLoader implements StandardLoader {
 	public StdSysPropLoader() {
 		unknownPropertyAProblem = false;
 	}
-	
+
 	@Override
-	public Map<?, ?> getMap() {
-		if (map != null) {
-			return map;
-		} else {
-			return System.getProperties();
-		}
+	public LoaderValues load(final PropertyConfigurationInternal runtimeDef, final ValidatedValuesWithContext existingValues) {
+		return null;
+	}
+
+	@Override
+	public LoaderValues load(final PropertyConfigurationInternal runtimeDef,
+			final LoaderEnvironment environment, final ValidatedValuesWithContext existingValues) {
+		return load(runtimeDef, existingValues, environment.getSystemProperties());
 	}
 	
 	@Override
